% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpEstimation.R
\name{fullEstimate}
\alias{fullEstimate}
\title{Full cell type profile estimation pipeline with plot outputs.}
\usage{
fullEstimate(exprData, genes, geneColName, groups, outDir,
  seekConsensus = FALSE, groupRotations = FALSE,
  outlierSampleRemove = FALSE, removeMinority = TRUE,
  geneTransform = function(x) {     homologene::mouse2human(x)$humanGene },
  comparisons = "all", pAdjMethod = stats::p.adjust.methods, PC = 1,
  estimateFile = NULL)
}
\arguments{
\item{exprData}{data.frame. Expression data. First collumns of the expression data should include gene names in the same format
as the ones specified in the marker gene lists. Any other non-expression related fields must not be of type 'double'}

\item{genes}{a named list containing marker gene lists of each cell type}

\item{geneColName}{character. name of the column containing the gene names in the expression file}

\item{groups}{a vector stating which groups each sample belongs to}

\item{outDir}{output directory for plots and tables}

\item{seekConsensus}{logical. If TRUE any gene with negative loadings in any of the groups individually will be
removed. Use if there is a high likelihood of gene regulation between the groups.}

\item{groupRotations}{logical. should the outputs include loadings calculated for individual genes}

\item{outlierSampleRemove}{logical. should the outlier samples be removed from the final output}

\item{removeMinority}{logical. If TRUE, decides which sign is the most common and removes genes from the minority sign.
Note that results will always be rotated in the positive direction Setting
seekConsensus to TRUE makes this irrelevant.}

\item{geneTransform}{a function that will be applied to the gene list. the default behavior is to change mouse genes
to human genes. set to NULL to keep the genes as they are}

\item{comparisons}{a 2 x n character matrix, just 'all' or NULL. Names of groups to compare when calculating p values
For each column the two groups indicated in the rows will be compared by wilcox.test.}

\item{pAdjMethod}{character. which method to use when adjusting p values for multiple testing correction}

\item{PC}{integer. which principal component to use when calculating cell type profile.}

\item{estimateFile}{name of the file that contains the final profile estimations}
}
\description{
Calculates cell type profiles in whole tissue datasets using marker genes provided. Fine tunes the
the calculation based on experimental groups if needed. This is only useful for quickly plotting profile estimations
of data with discrete experimental groups
}
\seealso{
\code{\link{mgpEstimate}}
}
