% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogboxFunctions.R
\name{geom_boxvio}
\alias{geom_boxvio}
\title{A violin plot overlayed by a boxplot}
\usage{
geom_boxvio(data = NULL, mapping = NULL)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \link[ggplot2]{fortify}. If not specified, must
be suppled in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If not
specified, must be suppled in each layer added to the plot.}
}
\description{
This function returns a list of ggplot elements that overlays a boxplot over
a violin plot.
}
