% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogboxFunctions.R
\docType{data}
\name{list}
\alias{list}
\title{Get individual components of a list output}
\format{An object of class \code{result} of length 1.}
\usage{
list
}
\description{
Separates a list into components. Usually used to split outputs
from functions that returns list. Taken from stackoverflow question http://stackoverflow.com/questions/1826519/function-returning-more-than-one-value
}
\examples{
list[a,b] = list(a=c(1,2,3), b= c(2,3,4))
list[c,] = list(c=c(1,2,3), d= c(2,3,4))
}
\keyword{datasets}
