% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGeneSelection.R
\name{markerCandidates}
\alias{markerCandidates}
\title{Write candidate marker genes to a file}
\usage{
markerCandidates(design, expression, outLoc, groupNames, regionNames = NULL,
  PMID = "PMID", rotate = NULL, cores = 1, debug = NULL,
  sampleName = "sampleName", replicates = "originalIndex",
  foldChangeThresh = 10, minimumExpression = 8, background = 6,
  regionHierarchy = NULL, geneID = "Gene.Symbol", seed = NULL)
}
\arguments{
\item{design}{data.frame. Metadata for the samples.}

\item{expression}{data.frame. Expression data for the samples. Gene names should be included as a column. Any other non expression
data not be of type \code{double}}

\item{outLoc}{Directory to save the candidate genes}

\item{groupNames}{column of the \code{design} with cell type names. If multiple columns are provided, selection will
be performed for each independently}

\item{regionNames}{column of the \code{design} with region names. Multiple regions can be listed separated by commas}

\item{PMID}{column of the \code{design} with pubmed identifiers. This is required to identify the samples coming
from the same study}

\item{rotate}{double. percentage of samples to be removed. 0.33 is used for the study}

\item{cores}{Number of cores to use for paralellization}

\item{debug}{Nothing to see here}

\item{sampleName}{column of the \code{design} with sample names matching the column names of the \code{expression} data}

\item{replicates}{column of the \code{design} that groups replicates}

\item{foldChangeThresh}{minimum fold change required for selection}

\item{minimumExpression}{minimum level of expression for a marker gene in its cell type}

\item{background}{level of expression that should be considered as background level}

\item{seed}{seed for random generation. if NULL will be set to random}

\item{regionHiearchy}{hiearchy of regions.}
}
\description{
Writes candidate marker genes to a file along with other required information (fold change and silhouette coefficient)
Selects candidates based on fold change to the median expression of other samples and a minimum expression in the
cell type
}

