% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionize.R
\name{memoReg}
\alias{memoReg}
\title{Memoised version of regionize function}
\usage{
memoReg(design, regionNames, groupNames, regionHierarchy = NULL)
}
\description{
This function is a \code{\link[memoise]{memoise}}d version of the \code{\link{regionize}} function
}

