% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{mgp_LesnickCroppedMeta}
\alias{mgp_LesnickCroppedMeta}
\title{Crop of Lesnick et al. dataset of Parkinson's disease patients and controls}
\format{An object of class \code{data.frame} with 25 rows and 2 columns.}
\usage{
mgp_LesnickCroppedMeta
}
\description{
This dataset is provided to be used as an example for marker gene estimation. Most of the metadata
columns are left behind. Full dataset is available in the analyiss pakcage from Mancarci et al.
}
\seealso{
\code{\link{mgp_LesnickCroppedExpression}}
}
\keyword{datasets}
