% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{mgp_LesnickParkinsonsExp}
\alias{mgp_LesnickParkinsonsExp}
\title{Crop of Lesnick et al. dataset of Parkinson's disease patients and controls}
\format{An object of class \code{data.frame} with 54675 rows and 29 columns.}
\usage{
mgp_LesnickParkinsonsExp
}
\description{
This dataset is provided to be used as an example for marker gene estimation.
}
\seealso{
\code{\link{mgp_LesnickCroppedMeta}}
}
\keyword{datasets}
