% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{mgp_sampleProfiles}
\alias{mgp_sampleProfiles}
\title{Mock dataset for cell type profiles}
\format{An object of class \code{data.frame} with 6 rows and 19 columns.}
\usage{
mgp_sampleProfiles
}
\description{
This is a simplified mock dataset that represents the minimal information
required for marker gene selection. Gene symbols or other form of gene
identifier to be outputted should be present in a column before the
expression data. If there are additional columns that are not the expression
data from samples they must not be of type double. In the Mancarci et al.
study, expression data from samples is microarray data from purified brain cell types.
Marker gene selection pipeline is designed to accomodate such microarray data that is on
log2 scale. To use any other form of data, it must be converted to log scale and parameters
for marker gene selection needs to be adjusted accordingly.
}
\seealso{
\code{\link{mpg_sampleProfilesMeta}}, \code{\link{mpg_sampleRegionHiearchy}}, input of \code{\link{mpg_markerCandidates}}
}
\keyword{datasets}
