% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{mgp_sampleProfilesMeta}
\alias{mgp_sampleProfilesMeta}
\title{Mock metadata for the mock dataset for cell type profiles}
\format{An object of class \code{data.frame} with 18 rows and 4 columns.}
\usage{
mgp_sampleProfilesMeta
}
\description{
This is a simplified mock dataset that represents the minimal information
required for marker gene selection. The metadata columns must include:
\itemize{
\item sample names that match the column names of profile
expression matrix ('sampleName' in this example)
 \item a column that indicates which samples are replicates of each other
 ('replicate' in this example)
\item a column that indicates which samples are from the same study
('PMID' in this example)
 \item a column that indicates which cell type the
sample represents. } Additionally a column can be added to represent the
brain region that the sample is extracted from ('region' in this example). If
multiple regions apply they can be added to the same column and be separated
with commas without spaces. Gene selection functions will consider regions
seperately and output results for each of them. If desired one can create a
region hierarchy (see \code{\link{mpg_sampleRegionHiearchy}}) in order to
indicate regions that are subsets of each other.
}
\seealso{
\code{\link{mpg_sampleProfiles}},
  \code{\link{mpg_sampleRegionHiearchy}}, input of
  \code{\link{mpg_markerCandidates}}
}
\keyword{datasets}

