% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostVariable.R
\name{mostVariable}
\alias{mostVariable}
\title{Select most variable probeset}
\usage{
mostVariable(x, genes = "Gene.Symbol", threshold = 6,
  threshFun = max)
}
\arguments{
\item{x}{Expression dataset to filter. A data frame. Gene names must be present as a column before the expression data}

\item{genes}{Character. Name of the column that lists gene symbols}

\item{threshold}{If filtering should be done the threshold of expression}

\item{threshFun}{If filtering should be done, function to apply to the gene row to compare with the threshold. By default genes are removed if maximum expression is below 6}
}
\description{
This function selects the most variable probeset each gene in an expression data. It also filters expression data of low expressed genes
}
