% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{mouseMarkerGenes}
\alias{mouseMarkerGenes}
\title{Brain cell type specific genes}
\format{An object of class \code{list} of length 15.}
\usage{
mouseMarkerGenes
}
\description{
A list of brain cell type specific genes selected from mouse pooled cell type
microarray data. Published in Mancarci et al. 2016.
}
\keyword{datasets}

