% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{mpg_sampleRegionHiearchy}
\alias{mpg_sampleRegionHiearchy}
\title{Mock hierarchy of regions}
\format{An object of class \code{list} of length 1.}
\usage{
mpg_sampleRegionHiearchy
}
\description{
This is a simplified mock hiearchy of regions. All represents whole tissue
while Region 1 and Region 2 are two subregions. If provided to gene selection
functions genes will be selected for All, Region 1 and Region 2 regions
seperately
}
\keyword{datasets}

