% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerGeneSelection.R
\name{pickMarkers}
\alias{pickMarkers}
\title{Pick marker genes out of candidate lists}
\usage{
pickMarkers(geneLoc, rotationThresh = 0.95, silhouette = 0.5,
  minSilhouette = 0, foldChange = 10, lilah = F)
}
\description{
Picks marker genes out of candidate lists. Behaves differently based on the number of columns the files have.
}
