% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpEstimation.R
\name{plotEstimates}
\alias{plotEstimates}
\title{Plot cell type profile estimates}
\usage{
plotEstimates(estimates, groups, plotNames, sigTest = wilcox.test,
  pAdjMethod = p.adjust.methods, comparisons = "all")
}
\arguments{
\item{groups}{A names list of groups $groups part pf the mgpEstimate function's output}

\item{plotNames}{a vector of strings. full address of the plots}

\item{sigTest}{function to be used in calculation of p values for comparisons}

\item{pAdjMethod}{character. which method to use when adjusting p values for multiple testing correction}

\item{comparisons}{a 2 by n matrix that includes group name pairs, indicating which groups should be compared when
calculating the p values. or "all" which will make comparison between all available groups by setting it to
combn(groups,2)}

\item{estimates.}{a named list of cell type estimates. $estimates part of the mgpEstimate function's output.}
}
\description{
A convenience function to plot the estimates
}

