% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{import}
\name{reexports}
\alias{\%<>\%}
\alias{\%>\%}
\alias{\%dorng\%}
\alias{list}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{doRNG}{\code{\link[doRNG]{\%dorng\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}}

  \item{ogbox}{\code{\link[ogbox]{list}}}
}}
\keyword{internal}

