% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionize.R
\name{regionize}
\alias{regionize}
\title{Create cell type annotations per region.}
\usage{
regionize(design, regionNames, groupNames, regionHierarchy = NULL)
}
\arguments{
\item{design}{A data.frame containing the design information}

\item{regionNames}{a string naming the collumn of the design table that indicates the region names}

\item{groupNames}{a string vector naming  the collumns of the design table that indicates cell type naming schemes.}

\item{regionHierarchy}{a list of lists that represents the region hierarchy.}
}
\value{
A list containing character vectors that annotates samples. If the sample will not be considered for the region
it will be represented as an NA, otherwise it'll have the cell type name indicated in the region
}
\description{
For each brain region, based on a defined hierarchy, it choses which samples to take in
}

