% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogboxFunctions.R
\name{replaceElement}
\alias{replaceElement}
\title{Replaces elements of a vector based on the dictonary provided}
\usage{
replaceElement(vector, dictionary = NULL, labels = NULL,
  NAreplace = NA)
}
\arguments{
\item{vector}{a vector}

\item{dictionary}{a named vector or just a vector if labels are provided}

\item{labels}{a character vector}

\item{NAreplace}{that to put instead of NAs}
}
\value{
A list with $newVector and $dictionary.
}
\description{
Replaces elements of a vector based on the dictonary provided
}
