% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogboxFunctions.R
\name{sepExpr}
\alias{sepExpr}
\title{Seprate expression data from gene annotations}
\usage{
sepExpr(data)
}
\arguments{
\item{data}{Expression data with gene annotations. First columns must inlcude
the gene annotations and they must not have data of type "double".}
}
\value{
A list of length 2. 1st element includes the gene annoation data, the
second element includes the gene expression data.
}
\description{
Separetes numberic component of gene expression data from non-numeric gene
annotations. Detects the first non-double column in the data frame and
separates the data into two parts
}
