% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogboxFunctions.R
\name{toColor}
\alias{toColor}
\title{Turn every member of Vector to a color from the palette}
\usage{
toColor(vector, palette = NULL, NAcolor = "white")
}
\arguments{
\item{vector}{a vector}

\item{NAcolor}{that to put instead of NAs}

\item{dictionary}{a named vector}
}
\description{
Turn every member of Vector to a color from the palette
}
