% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpEstimation.R
\name{groupRotations}
\alias{groupRotations}
\title{Calculates rotations based on each group}
\usage{
groupRotations(exprData, genes, geneColName, groups, outDir,
  geneTransform = function(x) {     homologene::mouse2human(x)$humanGene })
}
\arguments{
\item{exprData}{data.frame. Expression data. First collumns of the expression data should include gene names in the
same format as the ones specified in the marker gene lists. Any other non-expression related fields must not be of
type 'double'}

\item{genes}{a named list containing marker gene lists of each cell type}

\item{geneColName}{character. name of the column containing the gene names in the expression file}

\item{groups}{a vector stating which groups each sample belongs to}

\item{outDir}{if provided group rotations will be saved there}

\item{geneTransform}{a function that will be applied to the gene list. the default behavior is to change mouse genes
to human genes. set to NULL to keep the genes as they are}
}
\description{
Calculates rotations based on each group
}

