% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%<>\%}
\alias{reexports}
\alias{list}
\alias{reexports}
\alias{\%dorng\%}
\alias{reexports}
\alias{\%dopar\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{doRNG}{\code{\link[doRNG]{\%dorng\%}}}

  \item{foreach}{\code{\link[foreach]{\%dopar\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}}

  \item{ogbox}{\code{\link[ogbox]{list}}}
}}

